Created an exception: DataObjectGeneralException
in LP_data.calculate_constraints() changed command: all_vars = list(s.symbolic_variables)

in nonlinear_systems_datastructure.py:
- in __init__ changed command: self.symbolic_domain_of_parameters = self.create_symbolic_domain_of_parameters()

- changed create_symbolic_domain_of_parameters(self) back to the old code

- changed get_upper_bound_constraint_expr to verify_upper_bound_constraint, and then changed it back to the old code.

in nonlinear_systems.py:
-added the commands:
        res_LP = LP_data.LP_solve(lp_method, Verbose=verbose)  # first solve the LP
        LP_data.solutions.append(res_LP)

-added the commands:
        ubf_temp = data_obj.construct_UBF()
        data_obj.UBF = ubf_temp
        LP_data.A, LP_data.B = LP_data.calculate_constraints(data_obj, res['violation'], dreal_precision_upper_bound)

in parser_nonlinear_systems.py:
-set p, gridstep, timeout to be int.

added config.py to initialize file paths.

+++ control_system_abstractions/parser/parser_nonlinear_systems.py
@@ -152,12 +152,12 @@ def parse_nonlinear(line):
                     assert int(value) > 0
                     solver_options.update({key: int(value)})
                 elif key == 'gridstep':
-                    assert float(value) > 2
-                    solver_options.update({key: float(value)})
+                    assert int(value) > 2
+                    solver_options.update({key: int(value)})
                 elif key == 'dreal_precision':
                     solver_options.update({key: float(value)})
                 elif key == 'timeout':
-                    solver_options.update({key: float(value)})
+                    solver_options.update({key: int(value)})
                 elif key == 'opt_method':
                     assert value in ['revised simplex', 'simplex', 'interior-point']
                     solver_options.update({key: value})
